Add methods to retrieve collection items asynchronously
Analogous to the way events are raised the ChildDataNodeCollection and the DataValueCollection
and their viewer equivalents ViewerChildDataNodeCollection and ViewerDataValueCollection
now provide RequestItems() and RequestItemsAsync() methods to request the current set of items
in the collections. The callback passed to RequestItems() is executed in the same thread as an
event handler attached to the 'Changed' event. The callback passed to RequestItemsAsync() is
executed in the same thread as an event handler attached to the 'ChangedAsync' event. Therefore
these methods are ideally for fetching items that are in the collection and keep notified of
changes by the corresponding events.